Collapse duplicated FindParentDataGrid helpers into shared PerformanceMonitor.Ui
Hoist the two visual-tree-walk helpers FindParentDataGrid (owning grid of a
clicked MenuItem) and FindParentDataGridFromElement (walk up from any element)
into one shared home, PerformanceMonitor.Ui/DataGridHelpers.cs, and delete the
ten hand-duplicated copies across Lite and the Darling viewer.

The shared FindParentDataGrid takes the superset object form (Lite's
ContextMenuHelper signature), behaviorally identical to the MenuItem copies for
the MenuItem every caller passes, so all 19 caller files compile and behave
unchanged. ServerTab / ViewerServerTab are partial classes and using static is
per-file, so every partial that CALLS a helper got its own
using static PerformanceMonitor.Ui.DataGridHelpers; call sites are untouched.

The deprecated Full Dashboard has no copies and is not touched.

Verified: Lite, Darling viewer, and Dashboard all build with 0 errors; the
capability pins stay green (Lite 6/6, Darling 6/6).
