+++ Darling/PerformanceMonitor.Darling.Viewer/AlertsHistoryTab.xaml.cs
@@ -17,6 +17,7 @@ using System.Windows.Threading;
 using PerformanceMonitor.Common;
 using PerformanceMonitor.Notifications;
 using PerformanceMonitor.Ui;
+using static PerformanceMonitor.Ui.DataGridHelpers;
 
 namespace PerformanceMonitor.Darling.Viewer;
 
@@ -389,21 +390,6 @@ public partial class AlertsHistoryTab : UserControl
 
     private void ExportToCsv_Click(object sender, RoutedEventArgs e) =>
         DataGridExport.ExportToCsv(sender, "alert_history", ViewerExportSettings.CsvSeparator);
-
-    #endregion
-
-    #region Helpers
-
-    private static DataGrid? FindParentDataGrid(MenuItem menuItem)
-    {
-        var contextMenu = menuItem.Parent as ContextMenu;
-        var target = contextMenu?.PlacementTarget as FrameworkElement;
-        while (target != null && target is not DataGrid)
-        {
-            target = System.Windows.Media.VisualTreeHelper.GetParent(target) as FrameworkElement;
-        }
-        return target as DataGrid;
-    }
 
     #endregion
