+++ Darling/PerformanceMonitor.Darling.Viewer/ViewerServerTab.ActualPlan.cs
@@ -10,6 +10,7 @@ using System;
 using System.Threading;
 using System.Windows;
 using System.Windows.Controls;
+using static PerformanceMonitor.Ui.DataGridHelpers;
 
 namespace PerformanceMonitor.Darling.Viewer;
