+++ Darling/PerformanceMonitor.Darling.Viewer/ViewerServerTab.Blocking.cs
@@ -14,6 +14,7 @@ using System.Windows;
 using System.Windows.Controls;
 using PerformanceMonitor.Common;
 using PerformanceMonitor.Ui;
+using static PerformanceMonitor.Ui.DataGridHelpers;
 
 namespace PerformanceMonitor.Darling.Viewer;
