+++ Darling/PerformanceMonitor.Darling.Viewer/ViewerServerTab.Deadlock.cs
@@ -13,6 +13,7 @@ using System.Windows.Controls;
 using System.Windows.Input;
 using PerformanceMonitor.Common;
 using PerformanceMonitor.Ui;
+using static PerformanceMonitor.Ui.DataGridHelpers;
 
 namespace PerformanceMonitor.Darling.Viewer;
