+++ Lite/Controls/FinOpsTab.xaml.cs
@@ -21,6 +21,7 @@ using PerformanceMonitorLite.Models;
 using PerformanceMonitorLite.Helpers;
 using PerformanceMonitorLite.Services;
 using PerformanceMonitor.Ui;
+using static PerformanceMonitor.Ui.DataGridHelpers;
 using PerformanceMonitor.PlanAnalysis;
 
 namespace PerformanceMonitorLite.Controls;
@@ -1012,31 +1013,6 @@ public partial class FinOpsTab : UserControl
     {
         if (_filterPopup != null)
             _filterPopup.IsOpen = false;
-    }
-
-    private static DataGrid? FindParentDataGridFromElement(DependencyObject element)
-    {
-        var current = element;
-        while (current != null)
-        {
-            if (current is DataGrid dg)
-                return dg;
-            current = VisualTreeHelper.GetParent(current);
-        }
-        return null;
-    }
-
-    #endregion
-
-    #region Helpers
-
-    private static DataGrid? FindParentDataGrid(MenuItem menuItem)
-    {
-        var contextMenu = menuItem.Parent as ContextMenu;
-        var target = contextMenu?.PlacementTarget as FrameworkElement;
-        while (target != null && target is not DataGrid)
-            target = VisualTreeHelper.GetParent(target) as FrameworkElement;
-        return target as DataGrid;
     }
 
     #endregion
